requirements.txt
fastapi
uvicorn
google-genai
redis
pydantic